+++ src/tests/orga/views/integration/test_event.py
@@ -357,7 +357,7 @@ def test_event_history_query_count(
     client.force_login(user)
     ContentType.objects.clear_cache()
 
-    with django_assert_num_queries(14):
+    with django_assert_num_queries(15):
         response = client.get(event.orga_urls.history)
 
     assert response.status_code == 200